Split build logic utils into header and cc
